+++ relay-system/src/service/registry.rs
@@ -104,34 +104,24 @@ impl Inner {
         let task_id = TaskId::from(&service);
         let group = self.services.entry(task_id).or_default();
 
-        // Cleanup group, evicting all terminated services, while we're at it.
-        group.instances.retain(|s| !s.handle.is_finished());
-
-        let id = ServiceId {
-            task: task_id,
-            instance_id: group.next_instance_id,
-        };
-        group.next_instance_id += 1;
-
-        let future = ServiceMonitor::wrap(service.future);
+        // Services are allowed to process as much work as possible before yielding to other,
+        // lower priority tasks. We want to prioritize service backlogs over creating more work
+        // for these services.
+        let future = tokio::task::unconstrained(service.future);
+        let future = ServiceMonitor::wrap(future);
         let metrics = Arc::clone(future.metrics());
 
-        let jh = crate::runtime::spawn_in(handle, task_id, future);
-        let (sjh, sjhe) = crate::service::status::split(jh);
+        let task_handle = crate::runtime::spawn_in(handle, task_id, future);
+        let (status_handle, handle) = crate::service::status::split(task_handle);
 
-        let service = Service {
-            instance_id: id.instance_id,
-            metrics,
-            handle: sjh,
-        };
-        group.instances.push(service);
+        group.add(metrics, status_handle);
 
-        sjhe
+        handle
     }
 
     fn metrics(&self) -> impl Iterator<Item = (ServiceId, ServiceMetrics)> + '_ {
         self.services.iter().flat_map(|(task_id, group)| {
-            group.instances.iter().map(|service| {
+            group.iter().map(|service| {
                 let id = ServiceId {
                     task: *task_id,
                     instance_id: service.instance_id,
@@ -151,15 +141,63 @@ impl Inner {
     }
 }
 
+/// Logical grouping for all service instances of the same service.
+///
+/// A single service can be started multiple times, each individual
+/// instance of a specific service is tracked in this group.
+///
+/// The group keeps track of a unique per service identifier,
+/// which stays unique for the duration of the runtime.
+///
+/// It also holds a list of all currently alive service instances.
 #[derive(Debug, Default)]
 struct ServiceGroup {
+    /// Next unique per-service id.
+    ///
+    /// The next instance started for this group will be assigned the id
+    /// and the id is incremented in preparation for the following instance.
     next_instance_id: u32,
-    instances: Vec<Service>,
+    /// All currently alive service instances or instances that have stopped
+    /// but are not yet remove from the list.
+    instances: Vec<ServiceInstance>,
 }
 
+impl ServiceGroup {
+    /// Adds a started service to the service group.
+    pub fn add(&mut self, metrics: Arc<RawMetrics>, handle: ServiceStatusJoinHandle) {
+        // Cleanup the group, evicting all finished services, while we're at it.
+        self.instances.retain(|s| !s.handle.is_finished());
+
+        let instance_id = self.next_instance_id;
+        self.next_instance_id += 1;
+
+        let service = ServiceInstance {
+            instance_id,
+            metrics,
+            handle,
+        };
+
+        self.instances.push(service);
+    }
+
+    /// Returns an iterator over all currently alive services.
+    pub fn iter(&self) -> impl Iterator<Item = &ServiceInstance> {
+        self.instances.iter().filter(|s| !s.handle.is_finished())
+    }
+}
+
+/// Collection of metadata the registry tracks per service instance.
 #[derive(Debug)]
-struct Service {
+struct ServiceInstance {
+    /// The per service group unique id for this instance.
     instance_id: u32,
+    /// A raw handle for all metrics tracked for this instance.
+    ///
+    /// The handle gives raw access to all tracked metrics, these metrics
+    /// should be treated as **read-only**.
     metrics: Arc<RawMetrics>,
+    /// A handle to the service instance.
+    ///
+    /// The handle has information about the completion status of the service.
     handle: ServiceStatusJoinHandle,
 }